+++ customk/reviews/tests/conftest.py
@@ -1,6 +1,7 @@
 import pytest
+from users.models import User
+from reviews.models import Review, ReviewImage
 from rest_framework.test import APIClient
-
 from classes.models import Class
 
 
@@ -8,17 +9,36 @@ from classes.models import Class
 def api_client():
     return APIClient()
 
+@pytest.fixture
+def sample_user(api_client):
+    return User.objects.create_user(
+        email="testuser@example.com", password="testpassword"
+    )
+
 
 @pytest.fixture
 def sample_class():
-    """
-    Class 생성 샘플 코드입니다.
-    """
     return Class.objects.create(
         title="Sample Class",
         description="This is a sample class",
         max_person=10,
         require_person=5,
         price=50000,
-        address={"state": "Seoul", "city": "Gangnam", "street": "Teheran-ro"},
+        address="서울시 강남구 테헤란로",
+    )
+
+@pytest.fixture
+def review(sample_class, sample_user):
+    return Review.objects.create(
+        user=sample_user,
+        class_id=sample_class,
+        review="Test review content",
+        rating="4.5",
+    )
+
+
+@pytest.fixture
+def review_image(review):
+    return ReviewImage.objects.create(
+        review=review, image_url="http://example.com/image.jpg"
     )
